separate transforms into files
