+++ src/compile/data/assemble.ts
@@ -4,15 +4,18 @@ import {VgData} from '../../vega.schema';
 import {DataComponent} from './';
 import {AggregateNode} from './aggregate';
 import {BinNode} from './bin';
+import {CalculateNode} from './calculate';
 import {DataFlowNode, OutputNode} from './dataflow';
 import {FacetNode} from './facet';
+import {FilterNode} from './filter';
 import {ParseNode} from './formatparse';
+import {IdentifierNode} from './indentifier';
+import {LookupNode} from './lookup';
 import {NonPositiveFilterNode} from './nonpositivefilter';
 import {NullFilterNode} from './nullfilter';
 import {SourceNode} from './source';
 import {StackNode} from './stack';
 import {TimeUnitNode} from './timeunit';
-import {CalculateNode, FilterNode, IdentifierNode, LookupNode} from './transforms';
 
 /**
  * Print debug information for dataflow tree.
