+++ src/compile/data/calculate.ts
@@ -1,0 +1,31 @@
+import {CalculateTransform} from '../../transform';
+import {duplicate} from '../../util';
+import {VgFormulaTransform} from '../../vega.schema';
+import {DataFlowNode} from './dataflow';
+
+/**
+ * We don't know what a calculate node depends on so we should never move it beyond anything that produces fields.
+ */
+export class CalculateNode extends DataFlowNode {
+  public clone() {
+    return new CalculateNode(duplicate(this.transform));
+  }
+
+  constructor(private transform: CalculateTransform) {
+    super();
+  }
+
+  public producedFields() {
+    const out = {};
+    out[this.transform.as] = true;
+    return out;
+  }
+
+  public assemble(): VgFormulaTransform {
+    return {
+      type: 'formula',
+      expr: this.transform.calculate,
+      as: this.transform.as
+    };
+  }
+}
